requirements.txt
fastapi
uvicorn
langchain
streamlit
passlib[bcrypt]
python-jose
python-dotenv
langchain_core
langchain-community 
langchain-huggingface 
langchain-groq 
faiss-cpu 
sentence-transformers 
pypdf